src/main/webapp/app/entities/site-org-temp/delete/site-org-temp-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, fakeAsync, inject, tick } from '@angular/core/testing';
import { HttpResponse, provideHttpClient } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { SiteOrgTempService } from '../service/site-org-temp.service';

import { SiteOrgTempDeleteDialogComponent } from './site-org-temp-delete-dialog.component';

describe('SiteOrgTemp Management Delete Component', () => {
  let comp: SiteOrgTempDeleteDialogComponent;
  let fixture: ComponentFixture<SiteOrgTempDeleteDialogComponent>;
  let service: SiteOrgTempService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [SiteOrgTempDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(SiteOrgTempDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(SiteOrgTempDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(SiteOrgTempService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
